python/requirements.txt
numpy == 1.26.1
websocket-client
keras == 2.15.0
scikit-learn
tensorflow == 2.15.0
